PolicyManager: Move payload_size to int64_t.
Code style encourages signed types instead of unsigned types and the
value being exposed is already signed. This patch fixes that for the
payload_size variable and adapts the BoxedValue::ValuePrinter
implementations to use the int64_t and uint64_t types.

BUG=None
TEST=Unit tests updated

Reviewed-on: https://chromium-review.googlesource.com/198269
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>

+++ policy_manager/fake_updater_provider.h
@@ -41,7 +41,7 @@ class FakeUpdaterProvider : public UpdaterProvider {
     return &var_new_version_;
   }
 
-  virtual FakeVariable<size_t>* var_payload_size() override {
+  virtual FakeVariable<int64_t>* var_payload_size() override {
     return &var_payload_size_;
   }
 
@@ -76,7 +76,7 @@ class FakeUpdaterProvider : public UpdaterProvider {
   FakeVariable<double> var_progress_{"progress", kVariableModePoll};
   FakeVariable<Stage> var_stage_{"stage", kVariableModePoll};
   FakeVariable<std::string> var_new_version_{"new_version", kVariableModePoll};
-  FakeVariable<size_t> var_payload_size_{"payload_size", kVariableModePoll};
+  FakeVariable<int64_t> var_payload_size_{"payload_size", kVariableModePoll};
   FakeVariable<std::string> var_curr_channel_{
       "curr_channel", kVariableModePoll};
   FakeVariable<std::string> var_new_channel_{"new_channel", kVariableModePoll};
